\name{snpgarbage}
\alias{snpgarbage}
\alias{data}
\docType{data}
\title{Specimen 'snpgeno' object filled with garbage data}
\description{This is a pretty minimal \code{\link{snpgeno}} object with random contents, generated by the example code in \code{\link{snpgeno}}. The reasons for also including it as a data object, are (i) \emph{your} convenience, and (ii) so it can be used in the example of \code{\link{get_genotype_encoding}} (qv).
}
\usage{data( snpgarbage)
}
\format{An object of class \code{\link{snpgeno}}.
}
\author{Mark Bravington <email: markb1@summerinsouth.net>}
\keyword{data}
